REFACTOR `PdfTestCase` & `ExcelTestCase` to test directory root
