+++ tests/Feature/Excel/ExcelExportFromViewModelTest.php
@@ -4,6 +4,7 @@ namespace Sfneal\ViewExport\Tests\Feature\Excel;
 
 use Sfneal\ViewExport\Excel\ExcelExportService;
 use Sfneal\ViewExport\Tests\Assets\ViewModels\TestViewModel;
+use Sfneal\ViewExport\Tests\ExcelTestCase;
 
 class ExcelExportFromViewModelTest extends ExcelTestCase
 {
