+++ tests/Feature/Pdf/PdfExportFromUrlTest.php
@@ -3,6 +3,7 @@
 namespace Sfneal\ViewExport\Tests\Feature\Pdf;
 
 use Sfneal\ViewExport\Pdf\PdfExportService;
+use Sfneal\ViewExport\Tests\PdfTestCase;
 
 class PdfExportFromUrlTest extends PdfTestCase
 {
