+++ tests/PdfTestCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Sfneal\ViewExport\Tests\Feature\Pdf;
+namespace Sfneal\ViewExport\Tests;
 
 use Illuminate\Support\Facades\Bus;
 use Illuminate\Support\Facades\Queue;
@@ -8,7 +8,6 @@ use Illuminate\Support\Facades\Storage;
 use Sfneal\Helpers\Laravel\LaravelHelpers;
 use Sfneal\ViewExport\Pdf\Utils\PdfExporter;
 use Sfneal\ViewExport\Pdf\Utils\PdfRenderer;
-use Sfneal\ViewExport\Tests\TestCase;
 
 abstract class PdfTestCase extends TestCase
 {
